switched logicGraph over to generic "long"'s (from uint64_t's). Not sure if this is a good idea. Could improve performance on 32bit platforms.

+++ L-CAD/lcadengine/include/lcadengine/logicGraph.h
@@ -1,6 +1,6 @@
 #ifndef LOGICGRAPH_H_INCLUDED
 #define LOGICGRAPH_H_INCLUDED
-#include <stdint.h>
+
 #include <stdbool.h>
 
 #include "utils/fastlist.h"
@@ -8,19 +8,19 @@
 typedef enum {AND, OR, XOR, UNITY, RAND} gateInputType;
 
 struct {
-	uint64_t ID; /* Sequential ID of Gate */
+	unsigned long ID; /* Sequential ID of Gate */
 	bool state; /* Current Output state for this GLI */
-	uint8_t delay; /* Propication delay of this gate */
+	unsigned int delay; /* Propication delay of this gate */
 	gateInputType inputMode; /* AND OR etc.... */
 	bool inputNegate; /* N infront of the mode */
 } typedef genericLogicInterface;
 
 struct {
-	uint64_t ID; /* ID of this connection */
+	unsigned long ID; /* ID of this connection */
 	genericLogicInterface *srcEp; /* pointer to the GLI of the source */
 	genericLogicInterface *drnEp; /* pointer to the GLI of the drain */
-	uint64_t srcID; /* ID of the source */
-	uint64_t drnID; /* ID of the drain */
+	unsigned long srcID; /* ID of the source */
+	unsigned long drnID; /* ID of the drain */
 } typedef connection;
 
 struct s_graph;
@@ -29,17 +29,17 @@ typedef struct s_graph graph;
 graph *create_graph();
 void delete_graph(graph *ctx);
 
-uint64_t add_gli(graph *ctx, gateInputType type, bool nin, uint8_t delay);
-void remove_gli(graph *ctx, uint64_t ID);
+unsigned long add_gli(graph *ctx, gateInputType type, bool nin, unsigned int delay);
+void remove_gli(graph *ctx, unsigned long ID);
 
-uint64_t add_conn(graph *ctx, uint64_t Src, uint64_t Snk);
-void remove_conn(graph *ctx, uint64_t ID);
+unsigned long add_conn(graph *ctx, unsigned long Src, unsigned long Snk);
+void remove_conn(graph *ctx, unsigned long ID);
 
-genericLogicInterface *get_gli(graph *ctx, uint64_t ID);
+genericLogicInterface *get_gli(graph *ctx, unsigned long ID);
 
-connection *get_conn_by_id(graph *ctx, uint64_t ID);
-fastlist *get_conns_by_src(graph *ctx, uint64_t srcID);
-fastlist *get_conns_by_drn(graph *ctx, uint64_t drnID);
+connection *get_conn_by_id(graph *ctx, unsigned long ID);
+fastlist *get_conns_by_src(graph *ctx, unsigned long srcID);
+fastlist *get_conns_by_drn(graph *ctx, unsigned long drnID);
 
 unsigned long get_node_count(graph *ctx);
 #endif
